+++ packages/next/test/fixtures/00-app-dir-actions-experimental-streaming/index.test.js
@@ -15,7 +15,8 @@ function findActionId(page, runtime) {
       return actionId;
     }
   }
-  return null;
+
+  throw new Error("Couldn't find action ID");
 }
 
 function generateFormDataPayload(actionId) {
@@ -292,6 +293,80 @@ describe(`${__dirname.split(path.sep).pop()}`, () => {
       } else {
         expect(res.headers.get('x-edge-runtime')).toBe('1');
       }
+    });
+
+    it('should work when a rewrite greedy matches an action rewrite', async () => {
+      const targetPath = `${basePath}/static`;
+      const canonicalPath = `/greedy-rewrite/${basePath}/static`;
+      const actionId = findActionId(targetPath, runtime);
+
+      const res = await fetch(
+        `${ctx.deploymentUrl}${canonicalPath}`,
+        generateFormDataPayload(actionId)
+      );
+
+      expect(res.status).toEqual(200);
+      expect(res.headers.get('x-matched-path')).toBe(targetPath + '.action');
+      expect(res.headers.get('content-type')).toBe('text/x-component');
+      if (runtime === 'node') {
+        expect(res.headers.get('x-vercel-cache')).toBe('MISS');
+      } else {
+        expect(res.headers.get('x-edge-runtime')).toBe('1');
+      }
+    });
+
+    it('should work on the index route', async () => {
+      const canonicalPath = '/';
+      const actionId = findActionId('', 'node');
+
+      const res = await fetch(
+        `${ctx.deploymentUrl}${canonicalPath}`,
+        generateFormDataPayload(actionId)
+      );
+
+      expect(res.status).toEqual(200);
+      expect(res.headers.get('x-matched-path')).toBe('/index.action');
+      expect(res.headers.get('content-type')).toBe('text/x-component');
+      expect(res.headers.get('x-vercel-cache')).toBe('MISS');
+    });
+  });
+
+  describe('rewrite to index', () => {
+    it('should work when user has a rewrite to the index route', async () => {
+      const canonicalPath = '/rewritten-to-index';
+      const actionId = findActionId('', 'node');
+
+      const res = await fetch(
+        `${ctx.deploymentUrl}${canonicalPath}`,
+        generateFormDataPayload(actionId)
+      );
+
+      expect(res.status).toEqual(200);
+      expect(res.headers.get('x-matched-path')).toBe('/index.action');
+      expect(res.headers.get('content-type')).toBe('text/x-component');
+      expect(res.headers.get('x-vercel-cache')).toBe('MISS');
+    });
+
+    it('should work when entire path is rewritten', async () => {
+      const actionId = findActionId('/static', 'node');
+
+      const res = await fetch(ctx.deploymentUrl, {
+        method: 'POST',
+        body: JSON.stringify([1337]),
+        headers: {
+          'Content-Type': 'text/plain;charset=UTF-8',
+          'Next-Action': actionId,
+          'x-rewrite-me': '1',
+        },
+      });
+
+      expect(res.status).toEqual(200);
+      expect(res.headers.get('x-matched-path')).toBe('/index.action');
+      expect(res.headers.get('x-vercel-cache')).toBe('MISS');
+
+      const body = await res.text();
+      // The action incremented the provided count by 1
+      expect(body).toContain('1338');
     });
   });